requirements.prod.txt
Django>=4.2.1,<4.3
djangorestframework>=3.14.0,<3.15
pandas>=2.0.2,<2.1
python-dateutil>=2.8.2,<2.9
aioredis>=2.0.1,<2.1
psycopg
drf-yasg